requirements.txt
aiogram
python-dotenv
tabulate
aiosqlite
apscheduler
